feat: refactor response handling and update validation constraints in group management
