+++ src/main/java/com/nexaworks/rafiq/dto/response/common/Response.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.dto;
+package com.nexaworks.rafiq.dto.response.common;
 
 public record Response<T>(Boolean success, T data) {
 }
